Rename reporter to result2oscal
